requirements.txt
eikon
python-dotenv
refinitiv-data
pandas
numpy
matplotlib
PyYAML
pytest
pyarrow
